refactor mui boilerplate code into separate files
